Consolidate asDouble implementations, run perf test in run-benchmarks #479
