+++ pom.xml
@@ -470,6 +470,35 @@
                             </execution>
                         </executions>
                     </plugin>
+                    <plugin>
+                        <groupId>org.codehaus.mojo</groupId>
+                        <artifactId>exec-maven-plugin</artifactId>
+
+                        <configuration>
+                            <systemProperties>
+                                <systemProperty>
+                                    <key>jvm.resource.tracing</key>
+                                    <value>false</value>
+                                </systemProperty>
+                            </systemProperties>
+                            <classpathScope>test</classpathScope>
+                            <cleanupDaemonThreads>false</cleanupDaemonThreads>
+                        </configuration>
+                        <executions>
+                            <execution>
+                                <id>AsDoubleImplementationPerfJLBH</id>
+                                <phase>test</phase>
+                                <goals>
+                                    <goal>java</goal>
+                                </goals>
+                                <configuration>
+                                    <mainClass>
+                                        net.openhft.chronicle.core.AsDoubleImplementationPerfJLBH
+                                    </mainClass>
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
                 </plugins>
             </build>
         </profile>
