+++ src/main/java/net/openhft/chronicle/core/util/StringUtils.java
@@ -381,63 +381,6 @@ public final class StringUtils {
         return true;
     }
 
-    @Deprecated(/* For removal in x.25 */)
-    public static double asDouble(long value, int exp, boolean negative, int decimalPlaces) {
-        if (decimalPlaces > 0 && value < Long.MAX_VALUE / 2) {
-            if (value < Long.MAX_VALUE / (1L << 32)) {
-                exp -= 32;
-                value <<= 32;
-            }
-            if (value < Long.MAX_VALUE / (1L << 16)) {
-                exp -= 16;
-                value <<= 16;
-            }
-            if (value < Long.MAX_VALUE / (1L << 8)) {
-                exp -= 8;
-                value <<= 8;
-            }
-            if (value < Long.MAX_VALUE / (1L << 4)) {
-                exp -= 4;
-                value <<= 4;
-            }
-            if (value < Long.MAX_VALUE / (1L << 2)) {
-                exp -= 2;
-                value <<= 2;
-            }
-            if (value < Long.MAX_VALUE / (1L << 1)) {
-                exp -= 1;
-                value <<= 1;
-            }
-        }
-        for (; decimalPlaces > 0; decimalPlaces--) {
-            exp--;
-            long mod = value % 5;
-            value /= 5;
-            int modDiv = 1;
-            if (value < Long.MAX_VALUE / (1L << 4)) {
-                exp -= 4;
-                value <<= 4;
-                modDiv <<= 4;
-            }
-            if (value < Long.MAX_VALUE / (1L << 2)) {
-                exp -= 2;
-                value <<= 2;
-                modDiv <<= 2;
-            }
-            if (value < Long.MAX_VALUE / (1L << 1)) {
-                exp -= 1;
-                value <<= 1;
-                modDiv <<= 1;
-            }
-            if (decimalPlaces > 1)
-                value += modDiv * mod / 5;
-            else
-                value += (modDiv * mod + 4) / 5;
-        }
-        final double d = Math.scalb((double) value, exp);
-        return negative ? -d : d;
-    }
-
     @Nullable
     public static String toTitleCase(@Nullable String name) {
         if (name == null || name.isEmpty())
